generalize wire ops sacrificing commutativity for less code
