+++ plonk/code/src/arithmetizer/trace/mod.rs
@@ -342,7 +342,7 @@ mod tests {
         let rng = &mut rand::thread_rng();
         let [x, y] = Arithmetizer::build::<2>();
         let input_values = vec![1, 2];
-        let output_wires: &[Wire; 1] = &[3 * (x.clone() * x) + (y * 5) - 47];
+        let output_wires: &[Wire; 1] = &[(x.clone() * x) * 3 + (y * 5) - 47];
         // build circuit
 
         let circuit = output_wires[0].arith().borrow();
