+++ plonk/code/src/lib.rs
@@ -23,7 +23,7 @@ mod tests {
         let rng = &mut rand::thread_rng();
         let [x, y] = Arithmetizer::build();
         let input_values = vec![1, 2];
-        let output_wires = &[3 * (x.clone() * x) + (y * 5) - 47];
+        let output_wires = &[(x.clone() * x) * 3 + (y * 5) - 47];
         debug!("{}", Arithmetizer::to_string(&input_values, output_wires));
         let (x, w) = &Arithmetizer::to_circuit(rng, input_values, output_wires, None)?;
         debug!("{}", poly_evaluations_to_string(x, w));
